requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-redis==5.3.0
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.4
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.0.4
redis